Move DicePanel to battledisplay package
