+++ src/games/strategy/triplea/ui/BattlePanel.java
@@ -43,6 +43,7 @@ import games.strategy.triplea.delegate.dataObjects.CasualtyDetails;
 import games.strategy.triplea.delegate.dataObjects.CasualtyList;
 import games.strategy.triplea.delegate.dataObjects.FightBattleDetails;
 import games.strategy.triplea.ui.battledisplay.BattleDisplay;
+import games.strategy.triplea.ui.battledisplay.DicePanel;
 import games.strategy.ui.SwingAction;
 import games.strategy.ui.SwingComponents;
 import games.strategy.ui.Util;
